+++ src/main/java/in/koreatech/koin/global/reflection/ReflectionUtils.java
@@ -1,10 +1,12 @@
-package in.koreatech.koin.global.model;
+package in.koreatech.koin.global.reflection;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
 
 import org.springframework.aop.framework.AopProxyUtils;
 
+import in.koreatech.koin.global.reflection.exception.ClassNotFoundException;
+import in.koreatech.koin.global.reflection.exception.MethodNotFoundException;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
@@ -20,8 +22,11 @@ public class ReflectionUtils {
             A annotation = object.getDeclaredAnnotation(annotationType);
             return extractAnnotationValue(annotation, methodName);
         } catch (Exception e) {
-
-            throw new IllegalArgumentException();
+            throw new ClassNotFoundException(String.format(
+                "클래스가 어노테이션을 가지고 있지 않습니다. class: %s, annotation: %s",
+                proxy,
+                annotationType
+            ));
         }
     }
 
@@ -33,7 +38,11 @@ public class ReflectionUtils {
             Method method = annotation.annotationType().getMethod(methodName);
             return method.invoke(annotation);
         } catch (Exception e) {
-            throw new RuntimeException("어노테이션의 필드를 가져올 수 없습니다. " + e.getMessage());
+            throw new MethodNotFoundException(String.format(
+                "클래스에 메서드가 없습니다. class: %s, method: %s",
+                annotation,
+                methodName
+            ));
         }
     }
 }
